requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
Pillow>=10.0
PyMuPDF>=1.23
requests>=2.31
numpy>=1.26
opencv-python-headless>=4.8
gunicorn>=21.2
